+++ changelog.md
@@ -1,3 +1,1 @@
-- fixed sleep bar overlay being rendered wrong
-- removed better days marked as incompatible since people said it should work
-- 
+- fixed a forge override method not working